refactor: experiment with header files
